+++ mobile/android/android-components/components/service/firefox-accounts/src/test/java/mozilla/components/service/fxa/FxaAccountManagerTest.kt
@@ -233,7 +233,7 @@ class FxaAccountManagerTest {
         val accountStorage: AccountStorage = mock()
         val profile = Profile("testUid", "test@example.com", null, "Test Profile")
         val constellation: DeviceConstellation = mock()
-        val account = StatePersistenceTestableAccount(profile, constellation)
+        val account = StatePersistenceTestableAccount(profile, constellation, ableToRecoverFromAuthError = false)
 
         val accountObserver: AccountObserver = mock()
         val manager = TestableFxaAccountManager(
@@ -256,12 +256,14 @@ class FxaAccountManagerTest {
         assertNull(account.persistenceCallback)
 
         assertFalse(manager.accountNeedsReauth())
+        assertFalse(account.accessTokenErrorCalled)
         verify(accountObserver, never()).onAuthenticationProblems()
 
         manager.initAsync().await()
 
         assertTrue(manager.accountNeedsReauth())
         verify(accountObserver, times(1)).onAuthenticationProblems()
+        assertTrue(account.accessTokenErrorCalled)
     }
 
     @Test(expected = FxaPanicException::class)
@@ -351,14 +353,15 @@ class FxaAccountManagerTest {
         verify(accountStorage).write("test")
     }
 
-    class StatePersistenceTestableAccount(private val profile: Profile, private val constellation: DeviceConstellation) : OAuthAccount {
+    class StatePersistenceTestableAccount(private val profile: Profile, private val constellation: DeviceConstellation, val ableToRecoverFromAuthError: Boolean = false) : OAuthAccount {
         var persistenceCallback: StatePersistenceCallback? = null
+        var accessTokenErrorCalled = false
 
-        override fun beginOAuthFlowAsync(scopes: Array<String>, wantsKeys: Boolean): Deferred<String?> {
+        override fun beginOAuthFlowAsync(scopes: Set<String>, wantsKeys: Boolean): Deferred<String?> {
             return CompletableDeferred("auth://url")
         }
 
-        override fun beginPairingFlowAsync(pairingUrl: String, scopes: Array<String>): Deferred<String?> {
+        override fun beginPairingFlowAsync(pairingUrl: String, scopes: Set<String>): Deferred<String?> {
             return CompletableDeferred("auth://url")
         }
 
@@ -374,9 +377,16 @@ class FxaAccountManagerTest {
             return CompletableDeferred(true)
         }
 
-        override fun getAccessTokenAsync(singleScope: String): Deferred<AccessTokenInfo?> {
+        override fun getAccessTokenAsync(
+            singleScope: String
+        ): Deferred<AccessTokenInfo?> {
             fail()
-            return CompletableDeferred()
+            return CompletableDeferred(null)
+        }
+
+        override fun checkAuthorizationStatusAsync(singleScope: String): Deferred<Boolean?> {
+            accessTokenErrorCalled = true
+            return CompletableDeferred(ableToRecoverFromAuthError)
         }
 
         override fun getTokenServerEndpointURL(): String {
@@ -398,7 +408,10 @@ class FxaAccountManagerTest {
         }
 
         override fun close() {
-            fail()
+            // Only expect 'close' to be called if we can't recover from an auth error.
+            if (ableToRecoverFromAuthError) {
+                fail()
+            }
         }
     }
 
@@ -786,6 +799,9 @@ class FxaAccountManagerTest {
         verify(accountObserver, never()).onAuthenticationProblems()
         assertFalse(manager.accountNeedsReauth())
 
+        // Our recovery flow should attempt to hit this API. Model the "can't recover" condition by returning 'false'.
+        `when`(mockAccount.checkAuthorizationStatusAsync(eq("profile"))).thenReturn(CompletableDeferred(false))
+
         // At this point, we're logged in. Trigger a 401.
         authErrorRegistry.notifyObservers {
             runBlocking(this@runBlocking.coroutineContext) {
@@ -796,18 +812,61 @@ class FxaAccountManagerTest {
         verify(accountObserver, times(1)).onAuthenticationProblems()
         assertTrue(manager.accountNeedsReauth())
         assertEquals(mockAccount, manager.authenticatedAccount())
+        assertNull(manager.accountProfile())
 
         // Able to re-authenticate.
         reset(accountObserver)
         assertEquals("auth://url", manager.beginAuthenticationAsync().await())
         assertEquals(mockAccount, manager.authenticatedAccount())
-        assertEquals(profile, manager.accountProfile())
 
         manager.finishAuthenticationAsync("dummyCode", "dummyState").await()
 
         verify(accountObserver).onAuthenticated(mockAccount)
         verify(accountObserver, never()).onAuthenticationProblems()
         assertFalse(manager.accountNeedsReauth())
+        assertEquals(profile, manager.accountProfile())
+    }
+
+    @Test
+    fun `authentication issues are recoverable via checkAuthorizationState`() = runBlocking {
+        val mockAccount: OAuthAccount = mock()
+        val constellation: DeviceConstellation = mock()
+        val profile = Profile(uid = "testUID", avatar = null, email = "test@example.com", displayName = "test profile")
+        val accountStorage = mock<AccountStorage>()
+        val accountObserver: AccountObserver = mock()
+        val manager = prepareHappyAuthenticationFlow(mockAccount, profile, accountStorage, accountObserver, this.coroutineContext)
+
+        // We start off as logged-out, but the event won't be called (initial default state is assumed).
+        verify(accountObserver, never()).onLoggedOut()
+        verify(accountObserver, never()).onAuthenticated(any())
+
+        reset(accountObserver)
+        assertEquals("auth://url", manager.beginAuthenticationAsync().await())
+        assertNull(manager.authenticatedAccount())
+        assertNull(manager.accountProfile())
+
+        `when`(mockAccount.deviceConstellation()).thenReturn(constellation)
+        `when`(constellation.initDeviceAsync(any(), any(), any())).thenReturn(CompletableDeferred(true))
+
+        manager.finishAuthenticationAsync("dummyCode", "dummyState").await()
+
+        verify(accountObserver, never()).onAuthenticationProblems()
+        assertFalse(manager.accountNeedsReauth())
+
+        // Recovery flow will hit this API, and will recover if it returns 'true'.
+        `when`(mockAccount.checkAuthorizationStatusAsync(eq("profile"))).thenReturn(CompletableDeferred(true))
+
+        // At this point, we're logged in. Trigger a 401.
+        authErrorRegistry.notifyObservers {
+            runBlocking(this@runBlocking.coroutineContext) {
+                onAuthErrorAsync(AuthException(AuthExceptionType.UNAUTHORIZED, FxaUnauthorizedException("401"))).await()
+            }
+        }
+
+        // Since we've recovered, outside observers should not have witnessed the momentary problem state.
+        verify(accountObserver, never()).onAuthenticationProblems()
+        assertFalse(manager.accountNeedsReauth())
+        assertEquals(mockAccount, manager.authenticatedAccount())
     }
 
     @Test
@@ -876,7 +935,7 @@ class FxaAccountManagerTest {
     }
 
     @Test
-    fun `profile fetching flow hit an auth problem`() = runBlocking {
+    fun `profile fetching flow hit an unrecoverable auth problem`() = runBlocking {
         val accountStorage = mock<AccountStorage>()
         val mockAccount: OAuthAccount = mock()
         val constellation: DeviceConstellation = mock()
@@ -889,6 +948,9 @@ class FxaAccountManagerTest {
             authErrorRegistry.notifyObservers { onAuthErrorAsync(AuthException(AuthExceptionType.UNAUTHORIZED)) }
             CompletableDeferred(value = null)
         }
+
+        // Our recovery flow should attempt to hit this API. Model the "can't recover" condition by returning false.
+        `when`(mockAccount.checkAuthorizationStatusAsync(eq("profile"))).thenReturn(CompletableDeferred(false))
 
         `when`(mockAccount.beginOAuthFlowAsync(any(), anyBoolean())).thenReturn(CompletableDeferred("auth://url"))
         `when`(mockAccount.completeOAuthFlowAsync(anyString(), anyString())).thenReturn(CompletableDeferred(true))
@@ -913,6 +975,7 @@ class FxaAccountManagerTest {
         verify(accountObserver, never()).onLoggedOut()
         verify(accountObserver, never()).onAuthenticated(any())
         verify(accountObserver, never()).onAuthenticationProblems()
+        verify(mockAccount, never()).checkAuthorizationStatusAsync(any())
         assertFalse(manager.accountNeedsReauth())
 
         reset(accountObserver)
@@ -924,6 +987,136 @@ class FxaAccountManagerTest {
 
         assertTrue(manager.accountNeedsReauth())
         verify(accountObserver, times(1)).onAuthenticationProblems()
+        verify(mockAccount, times(1)).checkAuthorizationStatusAsync(eq("profile"))
+        Unit
+    }
+
+    @Test
+    fun `profile fetching flow hit an unrecoverable auth problem for which we can't determine a recovery state`() = runBlocking {
+        val accountStorage = mock<AccountStorage>()
+        val mockAccount: OAuthAccount = mock()
+        val constellation: DeviceConstellation = mock()
+
+        `when`(mockAccount.deviceConstellation()).thenReturn(constellation)
+        `when`(constellation.initDeviceAsync(any(), any(), any())).thenReturn(CompletableDeferred(true))
+
+        `when`(mockAccount.getProfileAsync(anyBoolean())).then {
+            // Hit an auth error.
+            authErrorRegistry.notifyObservers { onAuthErrorAsync(AuthException(AuthExceptionType.UNAUTHORIZED)) }
+            CompletableDeferred(value = null)
+        }
+
+        // Our recovery flow should attempt to hit this API. Model the "don't know what's up" condition by returning null.
+        `when`(mockAccount.checkAuthorizationStatusAsync(eq("profile"))).thenReturn(CompletableDeferred(value = null))
+
+        `when`(mockAccount.beginOAuthFlowAsync(any(), anyBoolean())).thenReturn(CompletableDeferred("auth://url"))
+        `when`(mockAccount.completeOAuthFlowAsync(anyString(), anyString())).thenReturn(CompletableDeferred(true))
+        // There's no account at the start.
+        `when`(accountStorage.read()).thenReturn(null)
+
+        val manager = TestableFxaAccountManager(
+                context,
+                Config.release("dummyId", "bad://url"),
+                arrayOf("profile", "test-scope"),
+                accountStorage, coroutineContext = this.coroutineContext
+        ) {
+            mockAccount
+        }
+
+        val accountObserver: AccountObserver = mock()
+
+        manager.register(accountObserver)
+        manager.initAsync().await()
+
+        // We start off as logged-out, but the event won't be called (initial default state is assumed).
+        verify(accountObserver, never()).onLoggedOut()
+        verify(accountObserver, never()).onAuthenticated(any())
+        verify(accountObserver, never()).onAuthenticationProblems()
+        verify(mockAccount, never()).checkAuthorizationStatusAsync(any())
+        assertFalse(manager.accountNeedsReauth())
+
+        reset(accountObserver)
+        assertEquals("auth://url", manager.beginAuthenticationAsync().await())
+        assertNull(manager.authenticatedAccount())
+        assertNull(manager.accountProfile())
+
+        manager.finishAuthenticationAsync("dummyCode", "dummyState").await()
+
+        assertTrue(manager.accountNeedsReauth())
+        verify(accountObserver, times(1)).onAuthenticationProblems()
+        verify(mockAccount, times(1)).checkAuthorizationStatusAsync(eq("profile"))
+        Unit
+    }
+
+    @Test
+    fun `profile fetching flow hit a recoverable auth problem`() = runBlocking {
+        val accountStorage = mock<AccountStorage>()
+        val mockAccount: OAuthAccount = mock()
+        val constellation: DeviceConstellation = mock()
+
+        `when`(mockAccount.deviceConstellation()).thenReturn(constellation)
+        `when`(constellation.initDeviceAsync(any(), any(), any())).thenReturn(CompletableDeferred(true))
+
+        val profile = Profile(
+                uid = "testUID", avatar = null, email = "test@example.com", displayName = "test profile")
+        var didFailProfileFetch = false
+        `when`(mockAccount.getProfileAsync(anyBoolean())).then {
+            // Hit an auth error, but only once. As we recover from it, we'll attempt to fetch a profile
+            // again. At that point, we'd like to succeed.
+            if (!didFailProfileFetch) {
+                didFailProfileFetch = true
+                authErrorRegistry.notifyObservers { onAuthErrorAsync(AuthException(AuthExceptionType.UNAUTHORIZED)) }
+                CompletableDeferred(value = null)
+            } else {
+                CompletableDeferred(profile)
+            }
+        }
+
+        // Recovery flow will hit this API, return a success.
+        `when`(mockAccount.checkAuthorizationStatusAsync(eq("profile"))).thenReturn(CompletableDeferred(true))
+
+        `when`(mockAccount.beginOAuthFlowAsync(any(), anyBoolean())).thenReturn(CompletableDeferred("auth://url"))
+        `when`(mockAccount.completeOAuthFlowAsync(anyString(), anyString())).thenReturn(CompletableDeferred(true))
+        // There's no account at the start.
+        `when`(accountStorage.read()).thenReturn(null)
+
+        val manager = TestableFxaAccountManager(
+                context,
+                Config.release("dummyId", "bad://url"),
+                arrayOf("profile", "test-scope"),
+                accountStorage, coroutineContext = this.coroutineContext
+        ) {
+            mockAccount
+        }
+
+        val accountObserver: AccountObserver = mock()
+
+        manager.register(accountObserver)
+        manager.initAsync().await()
+
+        // We start off as logged-out, but the event won't be called (initial default state is assumed).
+        verify(accountObserver, never()).onLoggedOut()
+        verify(accountObserver, never()).onAuthenticated(any())
+        verify(accountObserver, never()).onAuthenticationProblems()
+        verify(mockAccount, never()).checkAuthorizationStatusAsync(any())
+        assertFalse(manager.accountNeedsReauth())
+
+        reset(accountObserver)
+        assertEquals("auth://url", manager.beginAuthenticationAsync().await())
+        assertNull(manager.authenticatedAccount())
+        assertNull(manager.accountProfile())
+
+        manager.finishAuthenticationAsync("dummyCode", "dummyState").await()
+
+        assertFalse(manager.accountNeedsReauth())
+        assertEquals(mockAccount, manager.authenticatedAccount())
+        assertEquals(profile, manager.accountProfile())
+        verify(accountObserver, never()).onAuthenticationProblems()
+        // Once for the initial auth success, then once again after we recover from an auth problem.
+        verify(accountObserver, times(2)).onAuthenticated(mockAccount)
+        // Verify that we went through the recovery flow.
+        verify(mockAccount, times(1)).checkAuthorizationStatusAsync(eq("profile"))
+        Unit
     }
 
     @Test(expected = FxaPanicException::class)